% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdArma-package.R
\docType{package}
\name{svdArma-package}
\alias{svdArma}
\alias{svdArma-package}
\title{svdArma: Compute full SVD decomposition}
\description{
Compute full SVD decomposition.
}
\author{
\strong{Maintainer}: Alberto Quaini \email{alberto91quaini@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
