% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdR.R
\name{svdR}
\alias{svdR}
\title{Compute the full svd decomposition of a real matrix}
\usage{
svdR(C)
}
\arguments{
\item{C}{matrix of real numbers.}
}
\value{
list containing the matrix of left singular vectors,
the matrix of right singular vectors and the vector of nonzero
singular values.
}
\description{
Compute the full svd decomposition of a real matrix.
}
